Don't pack a NUL for symbols
TIL ruby has rb_intern2, which accepts a char* and size. This is the
right thing for unpack to call when it's creating a symbol.

This commit also revises the error testing for custom type unpacking.
The symbol tests don't need to deal with NUL-termination anymore. I did
add a test for each of the custom types to ensure that they correctly
raise an unpack error if the buffer is too short by one character.