Add a v1.2 compatibility helper
Mochilo::Compat_1_2 has all the same methods as Mochilo. It packs in a
way that Mochilo v1.2 can unpack, but it unpacks anything that Mochilo
v1.2 or v1.3 can pack. This allows applications to upgrade the unpacking
side of the code separately from upgrading the packing side of the code.